+++ incrowd/frontend/.gitignore
@@ -2,3 +2,5 @@ node_modules/
 bower_components/
 build/
 .tmp/
+src/lib/
+.awspublish*
