+++ incrowd/frontend/package.json
@@ -16,6 +16,7 @@
     "gulp-csso": "^1.0.0",
     "gulp-filter": "^2.0.2",
     "gulp-flatten": "*",
+    "gulp-if": "^1.2.5",
     "gulp-inject": "*",
     "gulp-jade-usemin": "*",
     "gulp-jshint": "*",
