+++ incrowd/incrowd/settings.py
@@ -129,7 +129,7 @@ if ENV == 'shippable':
     # Running in development, so use a local MySQL database.
     DATABASES = {
         'default': {
-            'ENGINE': 'django.db.backends.mysql',
+            'ENGINE': 'django.db.backends.sqlite3',
             'NAME': 'incrowd',
             'USER': 'shippable',
             'HOST': '172.17.42.1',  # Access MySQL on a Docker host
@@ -141,8 +141,8 @@ else:
     DATABASES = {
         'default': {
             'ENGINE': (os.environ.get('DB_BACKEND') or
-                       'django.db.backends.mysql'),
-            'NAME': os.environ.get('DB_NAME') or 'incrowd',
+                       'django.db.backends.sqlite3'),
+            'NAME': os.environ.get('DB_NAME') or os.path.join(BASE_DIR, 'db.sqlite3'),
             'USER': os.environ.get('DB_USER') or 'incrowd',
             'PASSWORD': os.environ.get('DB_PASSWORD') or 'incrowd',
             'HOST': os.environ.get('DB_ADDR') or '192.168.59.103',
